Change PLUGIN_DIR default to ${pkglibdir}/plugins
According to The GNU Coding Standards, dynamically loaded libraries
should go to ${pkglibdir}, but that's a Makefile variable, not a
configure script variable, which means it can't go into config.h, as
that is generated before 'make' is ever called.

So we use ${libdir}/${PACKAGE} instead, which should be the same thing
as ${pkglibdir} under normal circumstances.